move check on existence of primary alias from postgres to rust; use i64 for entity_ids

+++ crates/minerva/src/attribute_storage.rs
@@ -118,7 +118,7 @@ impl RawAttributeStore for AttributeStore {
 
         let entity_names: Vec<String> = rows.iter().map(|row| row.entity_name.clone()).collect();
 
-        let entity_ids: Vec<i32> = entity_mapping
+        let entity_ids: Vec<i64> = entity_mapping
             .names_to_entity_ids(tx, &self.entity_type, &entity_names)
             .await
             .map_err(AttributeStorageError::EntityMappingError)?;
